# Fix for select point from table in ec_nistp scalar_mul